# openapi/openapi.yaml
openapi: 3.0.3
info:
  title: listarr-go
  description: |
    Public listarr-go API. Inspired by fisherd80/Listarr (MIT).
    Safe defaults: loopback listen, apply opt-in, no private host defaults.
  version: 0.5.1
servers:
  - url: http://127.0.0.1:8787
    description: Local loopback (default)
paths:
  /:
    get:
      operationId: uiIndex
      summary: Operator web UI (no API key required to load shell)
      responses:
        "200":
          description: HTML
  /health:
    get:
      operationId: health
      responses:
        "200":
          description: OK
  /api/v1/ui/bootstrap:
    get:
      operationId: uiBootstrap
      summary: Embedded UI bootstrap (returns apiKey; keep listen on loopback)
      responses:
        "200":
          description: Bootstrap payload for the operator UI
          content:
            application/json:
              schema:
                type: object
                properties:
                  appName: { type: string }
                  version: { type: string }
                  instanceName: { type: string }
                  apiKey: { type: string }
  /api/v1/system/status:
    get:
      operationId: systemStatus
      security:
        - ApiKeyHeader: []
      responses:
        "200":
          description: Status (includes storeBackend, plexConfigured; no secrets)
  /api/v1/settings:
    get:
      operationId: getSettings
      summary: Operator settings (includes secrets for authenticated clients)
      security:
        - ApiKeyHeader: []
      responses:
        "200":
          description: Settings document
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Settings"
    put:
      operationId: putSettings
      summary: Update settings, persist to store, hot-reload runtime
      security:
        - ApiKeyHeader: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Settings"
      responses:
        "200":
          description: Saved settings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Settings"
        "400":
          description: Validation error
        "503":
          description: Store not configured
  /api/v1/activity:
    get:
      operationId: listActivity
      summary: Recent persisted sync runs
      security:
        - ApiKeyHeader: []
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 500
            default: 50
      responses:
        "200":
          description: Activity list
        "503":
          description: Store not configured
  /api/v1/arr/instances:
    get:
      operationId: listArrInstances
      summary: Named *arr instances (names/kinds only; no URLs)
      security:
        - ApiKeyHeader: []
      responses:
        "200":
          description: Instance list
  /api/v1/arr/test:
    post:
      operationId: testArrConnection
      summary: Probe an *arr URL/API key via /api/v3/system/status
      security:
        - ApiKeyHeader: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [kind, url, apiKey]
              properties:
                name: { type: string }
                kind:
                  type: string
                  enum: [radarr, sonarr]
                url:
                  type: string
                  example: http://127.0.0.1:7878
                apiKey: { type: string }
                authCookie:
                  type: string
                  description: Optional Cookie header for reverse-proxy auth
      responses:
        "200":
          description: Connection OK
        "400":
          description: Validation error
        "502":
          description: Upstream *arr unreachable or unauthorized
  /api/v1/arr/{name}/options:
    get:
      operationId: listArrOptions
      summary: Root folders and quality profiles on a named instance
      security:
        - ApiKeyHeader: []
      parameters:
        - name: name
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Target options
        "404":
          description: Unknown instance
        "502":
          description: Upstream *arr error
  /api/v1/arr/{name}/importlists:
    get:
      operationId: listArrImportLists
      summary: Import List configs on an instance (metadata)
      security:
        - ApiKeyHeader: []
      parameters:
        - name: name
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Import lists
        "404":
          description: Unknown instance
  /api/v1/plex/auth/pin:
    post:
      operationId: plexCreatePin
      summary: Start Plex PIN link (opens plex.tv auth URL)
      security:
        - ApiKeyHeader: []
      responses:
        "200":
          description: PIN created
  /api/v1/plex/auth/pin/{id}:
    get:
      operationId: plexPollPin
      summary: Poll Plex PIN; persists token when linked
      security:
        - ApiKeyHeader: []
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: linked true/false
  /api/v1/plex/auth:
    delete:
      operationId: plexUnlink
      summary: Clear stored Plex auth token
      security:
        - ApiKeyHeader: []
      responses:
        "200":
          description: Unlinked
  /api/v1/plex/test:
    post:
      operationId: plexTest
      summary: Probe Plex Media Server /identity
      security:
        - ApiKeyHeader: []
      responses:
        "200":
          description: Connection OK
  /api/v1/plex/libraries:
    get:
      operationId: plexLibraries
      summary: List Plex library folder paths for Sync root folders
      security:
        - ApiKeyHeader: []
      parameters:
        - name: mediaType
          in: query
          schema:
            type: string
            enum: [movie, tv]
      responses:
        "200":
          description: Library paths
  /api/v1/discover/movies:
    get:
      operationId: discoverMovies
      security:
        - ApiKeyHeader: []
      responses:
        "200":
          description: Discover results
        "503":
          description: TMDB not configured
  /api/v1/discover/tv:
    get:
      operationId: discoverTV
      security:
        - ApiKeyHeader: []
      responses:
        "200":
          description: Discover results
  /api/v1/sync/preview:
    post:
      operationId: syncPreview
      security:
        - ApiKeyHeader: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SyncRequest"
      responses:
        "200":
          description: Preview result
  /api/v1/sync/apply:
    post:
      operationId: syncApply
      security:
        - ApiKeyHeader: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SyncRequest"
      responses:
        "200":
          description: Apply result
        "403":
          description: Apply disabled
components:
  securitySchemes:
    ApiKeyHeader:
      type: apiKey
      in: header
      name: X-Api-Key
  schemas:
    Settings:
      type: object
      required: [apiKey, instanceName, torboxSearchPerHour]
      properties:
        apiKey: { type: string }
        instanceName: { type: string }
        applyEnabled: { type: boolean, description: "Inverse of safeMode (writes allowed)" }
        safeMode: { type: boolean, description: "Read-only when true (default)" }
        torboxSearchPerHour: { type: integer, minimum: 1 }
        tmdbApiKey: { type: string }
        plex:
          $ref: "#/components/schemas/PlexSettings"
        arrInstances:
          type: array
          items:
            $ref: "#/components/schemas/ArrInstanceSettings"
        updatedAt: { type: string, format: date-time }
    PlexSettings:
      type: object
      properties:
        serverUrl:
          type: string
          example: http://127.0.0.1:32400
        token: { type: string }
        clientIdentifier: { type: string }
        accountUsername: { type: string }
    ArrInstanceSettings:
      type: object
      required: [name, kind, url, apiKey]
      properties:
        name: { type: string }
        kind:
          type: string
          enum: [radarr, sonarr]
        url:
          type: string
          example: http://127.0.0.1:7878
        apiKey: { type: string }
        authCookie:
          type: string
          description: Optional Cookie header for reverse-proxy auth (e.g. oauth2-proxy)
    SyncRequest:
      type: object
      required: [source, mediaType, target]
      properties:
        source:
          type: string
          enum: [tmdb, arr-library]
        mediaType:
          type: string
          enum: [movie, tv]
        sourceInstance:
          type: string
          description: Required when source=arr-library
        sourceFilter:
          type: object
          properties:
            monitoredOnly: { type: boolean }
            tagIds:
              type: array
              items: { type: integer }
            pathContains: { type: string }
            requireAllTags: { type: boolean }
        tmdbIds:
          type: array
          items: { type: integer }
        discover:
          type: object
        maxItems:
          type: integer
          maximum: 2000
        target:
          $ref: "#/components/schemas/Target"
    Target:
      type: object
      required: [rootFolderPath, qualityProfileId]
      properties:
        instance:
          type: string
          description: Named registry instance (default radarr/sonarr)
        rootFolderPath:
          type: string
          example: /data/movies
        qualityProfileId:
          type: integer
          minimum: 1
        tags:
          type: array
          items: { type: integer }
        monitored: { type: boolean }
        searchOnAdd: { type: boolean }
        seasonFolder: { type: boolean }